requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.3
django-appconf==1.0.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-heroku==0.3.1
django-imagekit==4.0.2
django-imagekit-cropper==1.16
django-taggit==0.23.0
Faker==1.0.0
gunicorn==19.9.0
pilkit==2.0
Pillow==5.3.0
psycopg2==2.7.6.1
pystache==0.5.4
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
static==1.1.1
static3==0.7.0
text-unidecode==1.2
whitenoise==4.1.2
